+++ cpp/common/arrow/arrow_helpers.h
