+++ cpp/common/arrow/arrow_tensor_serde.h
@@ -16,18 +16,27 @@
 
 #pragma once
 
-#include <serving/metaspore.pb.h>
+#include <common/metaspore.pb.h>
 #include <common/types.h>
 
-#include <arrow/record_batch.h>
+#include <arrow/tensor.h>
 #include <fmt/format.h>
 
 namespace metaspore::serving {
 
-class ArrowRecordBatchSerde {
+class ArrowTensorSerde {
   public:
-    static result<std::shared_ptr<arrow::RecordBatch>>
-    deserialize_from(const std::string &name, const PredictRequest &request);
+    static result<std::shared_ptr<arrow::Tensor>>
+    deserialize_from(const std::string &name, PredictRequest &request);
+
+    static result<std::shared_ptr<arrow::Tensor>>
+    deserialize_from(const std::string &name, PredictReply &reply);
+
+    static status serialize_to(const std::string &name, const arrow::Tensor &tensor,
+                               PredictRequest &request);
+
+    static status serialize_to(const std::string &name, const arrow::Tensor &tensor,
+                               PredictReply &reply);
 };
 
 } // namespace metaspore::serving
