+++ cpp/tests/serving/feature_compute_exec_test.cpp
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-#include <serving/arrow_helpers.h>
+#include <common/arrow/arrow_helpers.h>
 #include <serving/feature_compute_exec.h>
 #include <serving/test_utils.h>
 #include <common/types.h>
