+++ cpp/tests/serving/sparse_ctr_model_test.cpp
@@ -15,7 +15,7 @@
 //
 
 #include <fmt/ranges.h>
-#include <serving/arrow_helpers.h>
+#include <common/arrow/arrow_helpers.h>
 #include <serving/feature_extraction_model_input.h>
 #include <serving/ort_model.h>
 #include <serving/tabular_model.h>
