+++ cpp/tests/serving/sparse_lookup_model_test.cpp
@@ -15,7 +15,7 @@
 //
 
 #include <arrow/api.h>
-#include <serving/arrow_helpers.h>
+#include <common/arrow/arrow_helpers.h>
 #include <serving/sparse_lookup_model.h>
 #include <serving/test_utils.h>
